relocated css colors
