+++ src/shared/flutter/colors.ts
@@ -624,3 +624,154 @@ export const flutterCupertinoColors: { [key: string]: string } = {
 	"tertiarySystemGroupedBackground.highContrastElevatedColor": "ffebebf0",
 	"white": "ffffffff",
 };
+
+export const flutterCSSColors: { [key: string]: string } = {
+	"aliceBlue": "fff0f8ff",
+	"antiqueWhite": "fffaebd7",
+	"aqua": "ff00ffff",
+	"aquamarine": "ff7fffd4",
+	"azure": "fff0ffff",
+	"beige": "fff5f5dc",
+	"bisque": "ffffe4c4",
+	"black": "ff000000",
+	"blanchedAlmond": "ffffebcd",
+	"blue": "ff0000ff",
+	"blueViolet": "ff8a2be2",
+	"brown": "ffa52a2a",
+	"burlyWood": "ffdeb887",
+	"cadetBlue": "ff5f9ea0",
+	"chartreuse": "ff7fff00",
+	"chocolate": "ffd2691e",
+	"coral": "ffff7f50",
+	"cornflowerBlue": "ff6495ed",
+	"cornsilk": "fffff8dc",
+	"crimson": "ffdc143c",
+	"cyan": "ff00ffff",
+	"darkBlue": "ff00008b",
+	"darkCyan": "ff008b8b",
+	"darkGoldenRod": "ffb8860b",
+	"darkGray": "ffa9a9a9",
+	"darkGreen": "ff006400",
+	"darkGrey": "ffa9a9a9",
+	"darkKhaki": "ffbdb76b",
+	"darkMagenta": "ff8b008b",
+	"darkOliveGreen": "ff556b2f",
+	"darkOrange": "ffff8c00",
+	"darkOrchid": "ff9932cc",
+	"darkRed": "ff8b0000",
+	"darkSalmon": "ffe9967a",
+	"darkSeaGreen": "ff8fbc8f",
+	"darkSlateBlue": "ff483d8b",
+	"darkSlateGray": "ff2f4f4f",
+	"darkSlateGrey": "ff2f4f4f",
+	"darkTurquoise": "ff00ced1",
+	"darkViolet": "ff9400d3",
+	"deepPink": "ffff1493",
+	"deepSkyBlue": "ff00bfff",
+	"dimGray": "ff696969",
+	"dimGrey": "ff696969",
+	"dodgerBlue": "ff1e90ff",
+	"fireBrick": "ffb22222",
+	"floralWhite": "fffffaf0",
+	"forestGreen": "ff228b22",
+	"fuchsia": "ffff00ff",
+	"gainsboro": "ffdcdcdc",
+	"ghostWhite": "fff8f8ff",
+	"gold": "ffffd700",
+	"goldenRod": "ffdaa520",
+	"gray": "ff808080",
+	"green": "ff008000",
+	"greenYellow": "ffadff2f",
+	"grey": "ff808080",
+	"honeyDew": "fff0fff0",
+	"hotPink": "ffff69b4",
+	"indianRed": "ffcd5c5c",
+	"indigo": "ff4b0082",
+	"ivory": "fffffff0",
+	"khaki": "fff0e68c",
+	"lavender": "ffe6e6fa",
+	"lavenderBlush": "fffff0f5",
+	"lawnGreen": "ff7cfc00",
+	"lemonChiffon": "fffffacd",
+	"lightBlue": "ffadd8e6",
+	"lightCoral": "fff08080",
+	"lightCyan": "ffe0ffff",
+	"lightGoldenRodYellow": "fffafad2",
+	"lightGray": "ffd3d3d3",
+	"lightGreen": "ff90ee90",
+	"lightGrey": "ffd3d3d3",
+	"lightPink": "ffffb6c1",
+	"lightSalmon": "ffffa07a",
+	"lightSeaGreen": "ff20b2aa",
+	"lightSkyBlue": "ff87cefa",
+	"lightSlateGray": "ff778899",
+	"lightSlateGrey": "ff778899",
+	"lightSteelBlue": "ffb0c4de",
+	"lightYellow": "ffffffe0",
+	"lime": "ff00ff00",
+	"limeGreen": "ff32cd32",
+	"linen": "fffaf0e6",
+	"magenta": "ffff00ff",
+	"maroon": "ff800000",
+	"mediumAquaMarine": "ff66cdaa",
+	"mediumBlue": "ff0000cd",
+	"mediumOrchid": "ffba55d3",
+	"mediumPurple": "ff9370db",
+	"mediumSeaGreen": "ff3cb371",
+	"mediumSlateBlue": "ff7b68ee",
+	"mediumSpringGreen": "ff00fa9a",
+	"mediumTurquoise": "ff48d1cc",
+	"mediumVioletRed": "ffc71585",
+	"midnightBlue": "ff191970",
+	"mintCream": "fff5fffa",
+	"mistyRose": "ffffe4e1",
+	"moccasin": "ffffe4b5",
+	"navajoWhite": "ffffdead",
+	"navy": "ff000080",
+	"oldLace": "fffdf5e6",
+	"olive": "ff808000",
+	"oliveDrab": "ff6b8e23",
+	"orange": "ffffa500",
+	"orangeRed": "ffff4500",
+	"orchid": "ffda70d6",
+	"paleGoldenRod": "ffeee8aa",
+	"paleGreen": "ff98fb98",
+	"paleTurquoise": "ffafeeee",
+	"paleVioletRed": "ffdb7093",
+	"papayaWhip": "ffffefd5",
+	"peachPuff": "ffffdab9",
+	"peru": "ffcd853f",
+	"pink": "ffffc0cb",
+	"plum": "ffdda0dd",
+	"powderBlue": "ffb0e0e6",
+	"purple": "ff800080",
+	"rebeccaPurple": "ff663399",
+	"red": "ffff0000",
+	"rosyBrown": "ffbc8f8f",
+	"royalBlue": "ff4169e1",
+	"saddleBrown": "ff8b4513",
+	"salmon": "fffa8072",
+	"sandyBrown": "fff4a460",
+	"seaGreen": "ff2e8b57",
+	"seaShell": "fffff5ee",
+	"sienna": "ffa0522d",
+	"silver": "ffc0c0c0",
+	"skyBlue": "ff87ceeb",
+	"slateBlue": "ff6a5acd",
+	"slateGray": "ff708090",
+	"slateGrey": "ff708090",
+	"snow": "fffffafa",
+	"springGreen": "ff00ff7f",
+	"steelBlue": "ff4682b4",
+	"tan": "ffd2b48c",
+	"teal": "ff008080",
+	"thistle": "ffd8bfd8",
+	"tomato": "ffff6347",
+	"turquoise": "ff40e0d0",
+	"violet": "ffee82ee",
+	"wheat": "fff5deb3",
+	"white": "ffffffff",
+	"whiteSmoke": "fff5f5f5",
+	"yellow": "ffffff00",
+	"yellowGreen": "ff9acd32",
+};
